Added hidden Eclipse files
These files are not only necessary to be able to import the project into Eclipse
IDE, but they also capture important information like the compiler compliance
(1.8) and classpath for the SQLite JDBC driver.

If you insist on not using Eclipse, make sure you are at least aware of these
settings.